data_record_tool/requirements.txt
pyserial==3.5
pandas==1.3.3
matplotlib==3.4.3
